fix: resolve merge conflicts from main
