--- frontend/src/app/dashboard/dashboardPage.tsx
@@ -1,12 +1,0 @@
-import Content from "@/components/dashboardContent";
-import Navbar from "@/components/navbar";
-
-//dashboard page contains the dashboard components
-export default function DashboardPage() {
-    return (
-        <div>
-            <Navbar />
-            <Content />
-        </div>
-    );
-}
