+++ frontend/src/app/dashboard/page.tsx
@@ -1,4 +1,4 @@
-import Dashboard from "@/component/dashboard/Dashboard";
+import Dashboard from "@/components/dashboard/Dashboard";
 
 export default function page() {
     return <Dashboard />;
