+++ frontend/src/components/dashboard/Dashboard.tsx
@@ -1,0 +1,30 @@
+"use client";
+
+import { AuthenticationContext } from "@/context/AuthContext";
+import { useContext } from "react";
+import Navbar from "../navbar";
+import Content from "./DashboardContent";
+
+export default function Dashboard() {
+  const authContext = useContext(AuthenticationContext);
+
+  if (!authContext) {
+    return <div>loading</div>;
+  }
+
+  const { isLoggedIn, user, logout } = authContext;
+
+  console.log("isLoggedIn: " + isLoggedIn);
+  console.log("User: " + user);
+
+  return (
+    <>
+      {}
+      <div>
+        <Navbar />
+        <div>User role: {user?.role}</div>
+        <Content />
+      </div>
+    </>
+  );
+}
