+++ frontend/src/components/dashboard/DashboardContent.tsx
@@ -1,0 +1,7 @@
+export default function Content() {
+    return (
+        <div>
+            <h1>Content goes here</h1>
+        </div>
+    );
+}
