--- frontend/src/components/dashboardContent.tsx
@@ -1,7 +1,0 @@
-export default function Content() {
-    return (
-        <div>
-            <h1>Content goes here</h1>
-        </div>
-    );
-}
